Move frequency-update validation to dingo.gw.frequency_updates
check_frequency_updates and the min/max validators move verbatim out of
gw_samplers.py to a neutral module: dingo_pipe imports them at INI
time, the legacy sampler setters keep using them, and the composed
data-prep cropping port will call them at construction -- they outlive
the legacy samplers. The module header documents the licensing rule the
code embodies: narrowing the frequency range is only permitted when the
network was trained with random strain cropping covering the requested
bounds.
